move heap dep to dev, move ff-scheduler to tests
plus prettier
